Move spatial fields into spatial package
